change extSymm to float
this will allow us to include chirality if desired (e.g. for C1 point group, we would have extSymm = 0.5, assuming, as is typical, that we have a racemic mixture)

+++ source/Molecule.py
@@ -195,7 +195,7 @@ class Molecule:
         if (line.split()[0].upper() != 'EXTSYM'):
            print 'Extsym keyword required'
            exit()
-        self.extSymm = int(line.split()[1])
+        self.extSymm = float(line.split()[1])
 
 #read electronic degeneracy
         line = readGeomFc.readMeaningfulLine(file)
